Clean up previous commit
