+++ mobile/src/com/cradle/iitc_mobile/IITC_WebChromeClient.java
@@ -2,6 +2,8 @@ package com.cradle.iitc_mobile;
 
 import android.webkit.ConsoleMessage;
 import android.webkit.GeolocationPermissions;
+import android.webkit.JsPromptResult;
+import android.webkit.JsResult;
 import android.webkit.WebChromeClient;
 import android.webkit.WebView;
 
@@ -51,5 +53,27 @@ public class IITC_WebChromeClient extends WebChromeClient {
             return true; // message was handled
 
         return super.onConsoleMessage(message);
+    }
+
+    @Override
+    public boolean onJsAlert(final WebView view, final String url, final String message, final JsResult result) {
+        return new IITC_JsDialogHelper(IITC_JsDialogHelper.ALERT, view, url, message, null, result).shouldInterrupt();
+    }
+
+    @Override
+    public boolean onJsBeforeUnload(final WebView view, final String url, final String message, final JsResult result) {
+        return new IITC_JsDialogHelper(IITC_JsDialogHelper.UNLOAD, view, url, message, null, result).shouldInterrupt();
+    }
+
+    @Override
+    public boolean onJsConfirm(final WebView view, final String url, final String message, final JsResult result) {
+        return new IITC_JsDialogHelper(IITC_JsDialogHelper.CONFIRM, view, url, message, null, result).shouldInterrupt();
+    }
+
+    @Override
+    public boolean onJsPrompt(final WebView view, final String url, final String message, final String defaultValue,
+            final JsPromptResult result) {
+        return new IITC_JsDialogHelper(IITC_JsDialogHelper.PROMPT, view, url, message, defaultValue, result)
+                .shouldInterrupt();
     }
 }
